[UPDATE] postgrates 통신 로직과 flirtinglist 랜더링 로직 서버사이드로 이동
