+++ src/types/realTimeType.ts
@@ -1,0 +1,7 @@
+import { RealtimePostgresChangesPayload } from '@supabase/supabase-js';
+
+export type SubscribeFlirtingListCallbackType = (
+  payload: RealtimePostgresChangesPayload<{
+    [key: string]: any;
+  }>
+) => void;
